Verification delete post and create post and category

+++ backend/src/posts.py
@@ -17,23 +17,21 @@ bp = Blueprint('posts', __name__, url_prefix='/')
 def showthreads(threadid, categoryid):
     cnx = get_db()
     cursor = cnx.cursor()
-    results = cursor.execute("SELECT title, content, timestamp, threadid, username, p.id FROM post as p LEFT JOIN user ON userid=user.id WHERE threadid = %s", (threadid,))
+    results = cursor.execute("SELECT title, content, timestamp, threadid, username, p.id, p.userid FROM post as p LEFT JOIN user ON userid=user.id WHERE threadid = %s", (threadid,))
     posts = cursor.fetchall() 
     click.echo(str(posts))
-    return render_template("post/index.html", posts=posts, categoryid=categoryid, threadid=threadid)
+    return render_template("post/index.html", posts=posts, categoryid=categoryid, threadid=threadid, isuser=g.user)
 
 @bp.route('/<categoryid>/<threadid>/', methods=['POST'])
 def create_newpost(threadid, categoryid):
     title=request.form['title']
     content=request.form['content']
     click.echo(title + content)
-    #TODO: check if both fields have text in them, Insert into database
     if title and content:
         cnx=get_db()
         cursor=cnx.cursor()
         timestamp=format(datetime.datetime.now())
-        #change userid to g.obj
-        userid=int(1)
+        userid=g.user.id
         cursor.execute("INSERT INTO `post` (`title`, `content`, `timestamp`, `userid`, `threadid`) VALUES (%s, %s, %s, %s, %s)", (title,content,timestamp,userid,int(threadid)))
         cnx.commit()
     return redirect("/"+categoryid+"/"+threadid+"/") 
@@ -41,11 +39,11 @@ def create_newpost(threadid, categoryid):
 
 @bp.route('/delete/<postid>', methods=['POST'])
 def delete_post(postid):
-	#TODO: check if the user is verified 
-	click.echo("delete " + postid)
-	cnx=get_db()
-	cursor=cnx.cursor()
-	cursor.execute("DELETE FROM post WHERE id = %s", (postid,))
-	cnx.commit()
-	return redirect("/")
+    if g.user:
+        click.echo("delete " + postid)
+        cnx=get_db()
+        cursor=cnx.cursor()
+        cursor.execute("DELETE FROM post WHERE id = %s", (postid,))
+        cnx.commit()
+        return redirect("/")
